Added file reading
You can now use a file to encode the intervals to be executed. The file is human-readable and the read function is quite robust to allow unmarked comments and ignore invalid data. Format is groups of 3 space-delimited integers separated by commas. Entries on the same line are stored on the same channel.