Added visualization and verbose file reading, updated README
Added VIsualize.jl to print a matrix of numbers for each time and channel to help see what a file is specifying. Updated Input.jl to include a verbose option that spits out nearly every decision it makes. Added a real README with some info, but it's not done or even up to date yet. Updated mynums.txt to better test resiliance of reading function. Updated PneuCore to send Python an array of values instead of a string.