Re-wrote project in Python
Everything is now in one Python file! I left the old files in for reference. I also switched to ANSI escape sequences for colors instead of whatever Julia does behind the scenes, and I updated the README to include slight changes in scope and instructions related to Python instead of Julia. Reading files now checks for negatives, and the text text file has a negative in it to test this.