Changed to module structure, improved I/O
The code is now divided into separate modules (like it was in Julia). You can now specify a gait file with command-line arguments or interactively in the program with cycle length and count, and the output is generally better-formatted and a bit more informative (with verbosity specified by the user at run time). Note that the 'ansicolor' pip3 package is now needed, but color will not show in IDLE, CMD, or PowerShell regardless.