Added Arduino PWM support, updated README
This is the minimum viable product; you can now control things! The Arduino code only works on a Nano, but it's designed to be edited easily for others. Currently assumes COM4 on Windows and /dev/ttyS4 otherwise. README now has correct script names and a few more lines about Arduinos.