Fixed PWM Frequency
The frequency was too high to let the solenoids actuate as required, so I switched to the Arduino code from the old project, since it sets registers to slow the PWM output. This means output.py had to change to interface with the old Arduino code. Note that the new code now requires an Arduino Mega 2560 so the registers are correct, so the code is no longer independent of Arduino version.